+++ lib/CalDAV/Plugin.php
@@ -577,6 +577,19 @@ class Plugin extends DAV\ServerPlugin {
 
         }
 
+        if ($node instanceof ICalendarObjectContainer && $depth === 0) {
+
+            if(strpos($this->server->httpRequest->getHeader('User-Agent'), 'MSFT-WP/') === 0) {
+                // Windows phone incorrectly supplied depth as 0, when it actually
+                // should have set depth to 1. We're implementing a workaround here
+                // to deal with this.
+                $depth = 1;
+            } else {
+                throw new BadRequest('A calendar-query REPORT on a calendar with a Depth: 0 is undefined. Set Depth to 1');
+            }
+
+        }
+
         // If we're dealing with a calendar, the calendar itself is responsible
         // for the calendar-query.
         if ($node instanceof ICalendarObjectContainer && $depth == 1) {